added methods to calculate expenses_total/Budget remaining_balance

+++ db/migrate/20200310140723_create_budgets.rb
@@ -2,7 +2,7 @@ class CreateBudgets < ActiveRecord::Migration[6.0]
   def change
     create_table :budgets do |t|
       t.string :title
-      t.decimal :bank
+      t.integer :bank
       t.integer :user_id
     end
   end
