+++ db/migrate/20200310170814_create_expenses.rb
@@ -2,7 +2,7 @@ class CreateExpenses < ActiveRecord::Migration[6.0]
   def change
     create_table :expenses do |t|
       t.string :name
-			t.float :cost
+			t.integer :cost
 			t.string :date
       t.belongs_to :budget
       t.belongs_to :user
